less hardcoded paths
