+++ Retrieval_BM25/bm25_only_retriever.py
@@ -12,7 +12,9 @@ from pathlib import Path
 from concurrent.futures import ThreadPoolExecutor
 import numpy as np
 
-MAIN_DATA_DIR = open(f"{os.getenv('HOME')}/data_dir").read().strip() if (lambda f: f and f.strip())(open(f"{os.getenv('HOME')}/data_dir").read()) else "/data/horse/ws/inbe405h-unarxive"
+from get_paths import get_main_data_dir
+
+MAIN_DATA_DIR = get_main_data_dir()
 
 try:
     from fast_llamaindex_retriever import FastLlamaIndexBM25Retriever
