+++ Retrieval_BM25/fast_llamaindex_retriever.py
@@ -3,7 +3,9 @@
 Fast LlamaIndex BM25 Retriever - Keeps index loaded in memory
 """
 
-MAIN_DATA_DIR = open(f"{os.getenv('HOME')}/data_dir").read().strip() if (lambda f: f and f.strip())(open(f"{os.getenv('HOME')}/data_dir").read()) else "/data/horse/ws/inbe405h-unarxive"
+from get_paths import get_main_data_dir
+
+MAIN_DATA_DIR = get_main_data_dir()
 
 import json
 import logging
