+++ run_basic_RAG.py
@@ -17,7 +17,9 @@ import re
 import random
 import string
 
-MAIN_DATA_DIR = open(f"{os.getenv('HOME')}/data_dir").read().strip() if (lambda f: f and f.strip())(open(f"{os.getenv('HOME')}/data_dir").read()) else "/data/horse/ws/inbe405h-unarxive"
+from get_paths import get_main_data_dir
+
+MAIN_DATA_DIR = get_main_data_dir()
 
 # Generate a unique ID for log filename
 def get_unique_log_filename():
