+++ start_backend_from_enterprise_cloud.sh
@@ -6,7 +6,9 @@ VENV_ACTIVATE=$VENV_DIR/bin/activate
 SCRIPT_DIR=$(dirname "$(realpath "$0")")
 
 USERNAME="squai-demo"
+USERNAME="s3811141"
 CLUSTER="capella"
+FALLBACK_CLUSTER="alpha"
 HPC_URL=""
 LOCAL_PORT="8000"
 
@@ -57,6 +59,8 @@ done
 if [[ -z "$HPC_URL" ]]; then
 	HPC_URL="login1.${CLUSTER}.hpc.tu-dresden.de"
 fi
+	
+    FALLBACK_HPC_URL="login1.${FALLBACK_CLUSTER}.hpc.tu-dresden.de"
 
 # Setup virtual environment if needed
 if [[ ! -d $VENV_DIR ]] || [[ ! -e $VENV_ACTIVATE ]]; then
@@ -75,5 +79,6 @@ bash "$SCRIPT_DIR/continous_hpc/enterprise_cloud/run" \
 	--hpc-job-name "squai_fastapi_backend" \
 	--username "$USERNAME" \
 	--sbatch_file_name "backend.sh" \
+    --fallback-system-url "$FALLBACK_HPC_URL" \
 	--copy \
 	--local-port "$LOCAL_PORT"
